html/semantics/embedded-content-0/the-canvas-element/security.drawImage.image.html

<!DOCTYPE html>
<title>Canvas test: security.drawImage.image</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.security.html">security</a>.<a href="index.security.drawImage.html">drawImage</a>.image</h1>
<p class="desc">drawImage of different-origin image makes the canvas origin-unclean</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.security.drawImage.image">security.drawImage.image</a>
<li><a href="../annotated-spec/canvas.html#testrefs.security.toDataURL">security.toDataURL</a>
<li><a href="../annotated-spec/canvas.html#testrefs.security.getImageData">security.getImageData</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("drawImage of different-origin image makes the canvas origin-unclean");
_addTest(function(canvas, ctx) {

ctx.drawImage(document.getElementById('yellow.png'), 0, 0);
assert_throws("SECURITY_ERR", function() { canvas.toDataURL(); });
assert_throws("SECURITY_ERR", function() { ctx.getImageData(0, 0, 1, 1); });


});
</script>
<img src="http://www2.w3c-test.org/html/tests/approved/images/yellow.png" id="yellow.png" class="resource">
